\name{cellmethy}
\alias{cellmethy}
\docType{data}
\title{
The Mean Methylation Values of Cell-specific Genes}
\description{
This file represents the mean methylation values of the cell-specific genes in 33 cell-lines.}
\usage{data(cellmethy)}
\format{
  A data frame with 2320 observations on the following 33 variables.
  \describe{
    \item{\code{HMEC}}{a numeric vector}
    \item{\code{MCF10A}}{a numeric vector}
    \item{\code{BT20}}{a numeric vector}
    \item{\code{BT474}}{a numeric vector}
    \item{\code{CAMA1}}{a numeric vector}
    \item{\code{EFM19}}{a numeric vector}
    \item{\code{EFM192A}}{a numeric vector}
    \item{\code{HCC1187}}{a numeric vector}
    \item{\code{HCC1419}}{a numeric vector}
    \item{\code{HCC1500}}{a numeric vector}
    \item{\code{HCC1569}}{a numeric vector}
    \item{\code{HCC1937}}{a numeric vector}
    \item{\code{HCC1954}}{a numeric vector}
    \item{\code{HCC38}}{a numeric vector}
    \item{\code{MCF7}}{a numeric vector}
    \item{\code{MDAMB175VII}}{a numeric vector}
    \item{\code{MDAMB231}}{a numeric vector}
    \item{\code{MDAMB361}}{a numeric vector}
    \item{\code{MDAMB415}}{a numeric vector}
    \item{\code{MDAMB436}}{a numeric vector}
    \item{\code{MDAMB453}}{a numeric vector}
    \item{\code{MDAMB468}}{a numeric vector}
    \item{\code{SKBR3}}{a numeric vector}
    \item{\code{SUM149PT}}{a numeric vector}
    \item{\code{SUM159PT}}{a numeric vector}
    \item{\code{T47D}}{a numeric vector}
    \item{\code{ZR751}}{a numeric vector}
    \item{\code{ZR7530}}{a numeric vector}
    \item{\code{SUM159}}{a numeric vector}
    \item{\code{HCC2157}}{a numeric vector}
    \item{\code{MDAMB134}}{a numeric vector}
    \item{\code{PMC42}}{a numeric vector}
    \item{\code{IPH926}}{a numeric vector}
  }
}
\examples{
# data(cellmethy)
## maybe str(cellmethy) ; plot(cellmethy) ...
}
\keyword{datasets}