\name{classification}
\alias{classification}
\docType{data}
\title{
The content of cell component and the subcluster of breast cancer samples}
\description{
This file contains a total of 672 breast cancer samples and provides the content of 33 cell component of each sample.
The subcluster of each sample is can also accessed, which is determined based on the content of cell component.
}
\usage{data(classification)}
\format{
  A data frame with 672 observations on the following 34 variables.
  \describe{
    \item{\code{subtype}}{a factor with levels \code{cluster1} \code{cluster2} \code{cluster3} \code{cluster4} \code{cluster5} \code{cluster6} \code{cluster7} \code{cluster8} \code{cluster9}}
    \item{\code{HMEC}}{a numeric vector}
    \item{\code{MCF10A}}{a numeric vector}
    \item{\code{BT20}}{a numeric vector}
    \item{\code{BT474}}{a numeric vector}
    \item{\code{CAMA1}}{a numeric vector}
    \item{\code{EFM19}}{a numeric vector}
    \item{\code{EFM192A}}{a numeric vector}
    \item{\code{HCC1187}}{a numeric vector}
    \item{\code{HCC1419}}{a numeric vector}
    \item{\code{HCC1500}}{a numeric vector}
    \item{\code{HCC1569}}{a numeric vector}
    \item{\code{HCC1937}}{a numeric vector}
    \item{\code{HCC1954}}{a numeric vector}
    \item{\code{HCC38}}{a numeric vector}
    \item{\code{MCF7}}{a numeric vector}
    \item{\code{MDAMB175VII}}{a numeric vector}
    \item{\code{MDAMB231}}{a numeric vector}
    \item{\code{MDAMB361}}{a numeric vector}
    \item{\code{MDAMB415}}{a numeric vector}
    \item{\code{MDAMB436}}{a numeric vector}
    \item{\code{MDAMB453}}{a numeric vector}
    \item{\code{MDAMB468}}{a numeric vector}
    \item{\code{SKBR3}}{a numeric vector}
    \item{\code{SUM149PT}}{a numeric vector}
    \item{\code{SUM159PT}}{a numeric vector}
    \item{\code{T47D}}{a numeric vector}
    \item{\code{ZR751}}{a numeric vector}
    \item{\code{ZR7530}}{a numeric vector}
    \item{\code{SUM159}}{a numeric vector}
    \item{\code{HCC2157}}{a numeric vector}
    \item{\code{MDAMB134}}{a numeric vector}
    \item{\code{PMC42}}{a numeric vector}
    \item{\code{IPH926}}{a numeric vector}
  }
}
\examples{
# data(classification)
}
\keyword{datasets}
