\name{tissuecg}
\alias{tissuecg}
\docType{data}
\title{
The Methylation Values of CpG Sites}
\description{
This file represents methylation values of 482420 CpG sites in two breast tissues, which are from 450K platform.}
\usage{data(tissuecg)}
\format{
  A data frame with 485577 observations on the following 2 variables.
  \describe{
    \item{\code{Tumor.33}}{a numeric vector}
    \item{\code{Tumor.34}}{a numeric vector}
  }
}

\examples{
# data(tissuecg)
## maybe str(tissuecg) ; plot(tissuecg) ...
}
\keyword{datasets}
