\name{tissuegene}
\alias{tissuegene}
\docType{data}
\title{
The Methylation Values of Cell-specific Genes
}
\description{
The methylation values of specific genes in two breast tissues, which are obtained from the extractMethy function.
}
\usage{data(tissuegene)}
\format{
  A data frame with 2320 observations on the following 2 variables.
  \describe{
    \item{\code{Tumor.33}}{a numeric vector}
    \item{\code{Tumor.34}}{a numeric vector}
  }
}
\examples{
# data(tissuegene)
}
\keyword{datasets}
